replace fibers with deasync
