+++ packages/wdio-sync/src/index.js
@@ -1,83 +1,27 @@
-import Fiber from './fibers'
+import deasync from 'deasync'
 
-import executeHooksWithArgs from './executeHooksWithArgs'
-import runFnInFiberContext from './runFnInFiberContext'
-import wrapCommand from './wrapCommand'
-
-import { STACKTRACE_FILTER_FN } from './constants'
-const defaultRetries = { attempts: 0, limit: 0 }
-
-/**
- * execute test or hook synchronously
- *
- * @param  {Function} fn         spec or hook method
- * @param  {Number}   retries    { limit: number, attempts: number }
- * @param  {Array}    args       arguments passed to hook
- * @return {Promise}             that gets resolved once test/hook is done or was retried enough
- */
-const executeSync = async function (fn, retries = defaultRetries, args = []) {
-    /**
-     * User can also use the `@wdio/sync` package directly to run commands
-     * synchronously in standalone mode. In this case we neither have
-     * `global.browser` nor `this`
-     */
-    if (global.browser) {
-        delete global.browser._NOT_FIBER
-    }
-    if (this) {
-        this.wdioRetries = retries.attempts
+export default function waitForPromise (promise) {
+    if (!(promise instanceof Promise)) {
+        return promise
     }
 
-    try {
-        global._HAS_FIBER_CONTEXT = true
-        let res = fn.apply(this, args)
-        global._HAS_FIBER_CONTEXT = false
+    let error
+    let result
+    let isDone = false
 
-        /**
-         * sometimes function result is Promise,
-         * we need to await result before proceeding
-         */
-        if (res instanceof Promise) {
-            return await res
-        }
-
-        return res
-    } catch (e) {
-        if (retries.limit > retries.attempts) {
-            retries.attempts++
-            return await executeSync.call(this, fn, retries, args)
-        }
-
-        /**
-         * no need to modify stack if no stack available
-         */
-        if (!e.stack) {
-            return Promise.reject(e)
-        }
-
-        e.stack = e.stack.split('\n').filter(STACKTRACE_FILTER_FN).join('\n')
-        return Promise.reject(e)
-    }
-}
-
-/**
- * run hook or spec via executeSync
- */
-function runSync (fn, repeatTest = 0, args = []) {
-    return (resolve, reject) =>
-        Fiber(() => executeSync.call(this, fn, repeatTest, args).then(resolve, reject)).run()
-}
-
-export {
-    executeHooksWithArgs,
-    wrapCommand,
-    runFnInFiberContext,
-    executeSync,
-    runSync,
-}
-
-export default function sync (testFn) {
-    return new Promise((resolve, reject) => {
-        return runSync(testFn)(resolve, reject)
+    promise.then((res) => {
+        result = res
+        isDone = true
+    }, (err) => {
+        error = err
+        isDone = true
     })
+
+    deasync.loopWhile(() => !isDone)
+
+    if (error) {
+        throw error
+    }
+
+    return result
 }
