+++ packages/wdio-sync/tests/__mocks__/deasync.js
@@ -1,0 +1,3 @@
+export default {
+    loopWhile: (fn) => fn()
+}
