--- packages/wdio-sync/tests/__mocks__/errors.js
@@ -1,1 +1,0 @@
-export const anotherError = new Error('AnotherError')
