--- packages/wdio-sync/tests/__mocks__/fibers.js
@@ -1,5 +1,0 @@
-const FiberMock = jest.fn().mockImplementation((fn) => {
-    return { run: fn }
-})
-
-module.exports = FiberMock
