--- packages/wdio-sync/tests/__mocks__/fibers/future.js
@@ -1,2 +1,0 @@
-const Future = jest.requireActual('fibers/future')
-module.exports = Future
