+++ packages/wdio-utils/src/shim.js
@@ -3,8 +3,6 @@ import logger from '@wdio/logger'
 const log = logger('@wdio/utils:shim')
 
 let inCommandHook = false
-let hasWdioSyncSupport = false
-let runSync = null
 
 let executeHooksWithArgs = async function executeHooksWithArgsShim (hooks, args) {
     /**
@@ -48,99 +46,49 @@ let executeHooksWithArgs = async function executeHooksWithArgsShim (hooks, args)
     return Promise.all(hooks)
 }
 
-let runFnInFiberContext = function (fn) {
-    return function (...args) {
-        return Promise.resolve(fn.apply(this, args))
+function wrapCommandFactory (syncWrapper) {
+    return function wrapCommand (commandName, fn) {
+        async function wrapCommandFn (...args) {
+            const beforeHookArgs = [commandName, args]
+            if (!inCommandHook && this.options.beforeCommand) {
+                inCommandHook = true
+                await executeHooksWithArgs.call(this, this.options.beforeCommand, beforeHookArgs)
+                inCommandHook = false
+            }
+
+            let commandResult
+            let commandError
+            try {
+                commandResult = await fn.apply(this, args)
+            } catch (err) {
+                commandError = err
+            }
+
+            if (!inCommandHook && this.options.afterCommand) {
+                inCommandHook = true
+                const afterHookArgs = [...beforeHookArgs, commandResult, commandError]
+                await executeHooksWithArgs.call(this, this.options.afterCommand, afterHookArgs)
+                inCommandHook = false
+            }
+
+            if (commandError) {
+                throw commandError
+            }
+
+            return commandResult
+        }
+
+        if (typeof syncWrapper !== 'function') {
+            return wrapCommandFn
+        }
+
+        return function callFn (...args) {
+            return syncWrapper(wrapCommandFn.call(this, ...args))
+        }
     }
 }
 
-let wrapCommand = function wrapCommand (commandName, fn) {
-    return async function wrapCommandFn (...args) {
-        const beforeHookArgs = [commandName, args]
-        if (!inCommandHook && this.options.beforeCommand) {
-            inCommandHook = true
-            await executeHooksWithArgs.call(this, this.options.beforeCommand, beforeHookArgs)
-            inCommandHook = false
-        }
-
-        let commandResult
-        let commandError
-        try {
-            commandResult = await fn.apply(this, args)
-        } catch (err) {
-            commandError = err
-        }
-
-        if (!inCommandHook && this.options.afterCommand) {
-            inCommandHook = true
-            const afterHookArgs = [...beforeHookArgs, commandResult, commandError]
-            await executeHooksWithArgs.call(this, this.options.afterCommand, afterHookArgs)
-            inCommandHook = false
-        }
-
-        if (commandError) {
-            throw commandError
-        }
-
-        return commandResult
-    }
-}
-
-/**
- * execute test or hook synchronously
- *
- * @param  {Function} fn         spec or hook method
- * @param  {Number}   retries    { limit: number, attempts: number }
- * @param  {Array}    args       arguments passed to hook
- * @return {Promise}             that gets resolved once test/hook is done or was retried enough
- */
-let executeSync = async function (fn, retries, args = []) {
-    this.wdioRetries = retries.attempts
-
-    try {
-        let res = fn.apply(this, args)
-
-        /**
-         * sometimes function result is Promise,
-         * we need to await result before proceeding
-         */
-        if (res instanceof Promise) {
-            return await res
-        }
-
-        return res
-    } catch (e) {
-        if (retries.limit > retries.attempts) {
-            retries.attempts++
-            return await executeSync.call(this, fn, retries, args)
-        }
-
-        return Promise.reject(e)
-    }
-}
-
-/**
- * execute test or hook asynchronously
- *
- * @param  {Function} fn         spec or hook method
- * @param  {object}   retries    { limit: number, attempts: number }
- * @param  {Array}    args       arguments passed to hook
- * @return {Promise}             that gets resolved once test/hook is done or was retried enough
- */
-const executeAsync = async function (fn, retries, args = []) {
-    this.wdioRetries = retries.attempts
-
-    try {
-        return await fn.apply(this, args)
-    } catch (e) {
-        if (retries.limit > retries.attempts) {
-            retries.attempts++
-            return await executeAsync.call(this, fn, retries, args)
-        }
-
-        throw e
-    }
-}
+let waitForPromise = null
 
 /**
  * shim to make sure that we only wrap commands if wdio-sync is installed as dependency
@@ -154,24 +102,15 @@ try {
     /* istanbul ignore if */
     if (!process.env.WDIO_NO_SYNC_SUPPORT) {
         // eslint-disable-next-line import/no-unresolved
-        const wdioSync = require('@wdio/sync')
-        hasWdioSyncSupport = true
-        runFnInFiberContext = wdioSync.runFnInFiberContext
-        wrapCommand = wdioSync.wrapCommand
-        executeHooksWithArgs = wdioSync.executeHooksWithArgs
-        executeSync = wdioSync.executeSync
-        runSync = wdioSync.runSync
+        waitForPromise = require('@wdio/sync').default
     }
 } catch {
     // do nothing
 }
 
+const wrapCommand = wrapCommandFactory(waitForPromise)
+
 export {
     executeHooksWithArgs,
-    runFnInFiberContext,
-    wrapCommand,
-    hasWdioSyncSupport,
-    executeSync,
-    executeAsync,
-    runSync
+    wrapCommand
 }
